bug fix
I think it should be accurate now. Of note, the total number of reads again doesn't match the sum of column 7 of the plus and minus bed files, because there are some genes with 0 mappability that get NA value in the summary table, yet they somehow have some reads. To compare numbers between the bed files and the final table, from the bed files select lines that are not "intergenic" and that have mappability > 0.